Cortx 32236 te rvf multi part with versioning scenarios part2 (#1683)
* Automated: TEST 41288, TEST 41289 and TEST 41290.


* Fixed: PEP 8: E127 continuation line over-indented for visual indent and
pep257 reported issues.


* Fixed codacy issues from multipart libs.


* Addressed & incorporated review comments.
