+++ libs/s3/s3_multipart.py
@@ -20,11 +20,15 @@
 
 """Python Library using boto3 module to perform multipart Operations."""
 
+import logging
 import os
 import sys
-import logging
 import threading
+from typing import Union
+
 from boto3.s3.transfer import TransferConfig
+
+from commons.constants import ERR_MSG
 from libs.s3.s3_core_lib import S3Lib
 
 LOGGER = logging.getLogger(__name__)
@@ -33,11 +37,8 @@ LOGGER = logging.getLogger(__name__)
 class Multipart(S3Lib):
     """Class containing methods to implement multipart functionality."""
 
-    def create_multipart_upload(self,
-                                bucket_name: str = None,
-                                obj_name: str = None,
-                                m_key: str = None,
-                                m_value: str = None) -> dict:
+    def create_multipart_upload(self, bucket_name: str = None, obj_name: str = None,
+                                m_key: str = None, m_value: str = None) -> dict:
         """
         Request to initiate a multipart upload.
 
@@ -51,17 +52,13 @@ class Multipart(S3Lib):
             response = self.s3_client.create_multipart_upload(
                 Bucket=bucket_name, Key=obj_name, Metadata={m_key: m_value})
         else:
-            response = self.s3_client.create_multipart_upload(
-                Bucket=bucket_name, Key=obj_name)
+            response = self.s3_client.create_multipart_upload(Bucket=bucket_name, Key=obj_name)
         LOGGER.debug("Response: %s", str(response))
 
         return response
 
-    def upload_part(self,
-                    body: str = None,
-                    bucket_name: str = None,
-                    object_name: str = None,
-                    **kwargs) -> dict:
+    def upload_part(self, body: Union[str, bytes] = None, bucket_name: str = None,
+                    object_name: str = None, **kwargs) -> dict:
         """
         Upload parts of a specific multipart upload.
 
@@ -76,48 +73,40 @@ class Multipart(S3Lib):
         content_md5 = kwargs.get("content_md5", None)
         if content_md5:
             response = self.s3_client.upload_part(
-                Body=body, Bucket=bucket_name, Key=object_name,
-                UploadId=upload_id, PartNumber=part_number,
-                ContentMD5=content_md5)
+                Body=body, Bucket=bucket_name, Key=object_name, UploadId=upload_id,
+                PartNumber=part_number, ContentMD5=content_md5)
         else:
             response = self.s3_client.upload_part(
-                Body=body, Bucket=bucket_name, Key=object_name,
-                UploadId=upload_id, PartNumber=part_number)
+                Body=body, Bucket=bucket_name, Key=object_name, UploadId=upload_id,
+                PartNumber=part_number)
         logging.debug(response)
 
         return response
 
-    def list_parts(
-            self,
-            mpu_id: str = None,
-            bucket: str = None,
-            object_name: str = None,
-            part_number_marker: int = 0) -> dict:
+    def list_parts(self, mpu_id: str = None, bucket_name: str = None,
+                   object_name: str = None, **kwargs) -> dict:
         """
         list parts of a specific multipart upload.
         :param mpu_id: Multipart upload ID
-        :param bucket: Name of the bucket
+        :param bucket_name: Name of the bucket
         :param object_name: Name of the object
-        :param part_number_marker: next part number in case parts greater than 1000.
+        :keyword part_number_marker: next part number in case parts greater than 1000.
         :return: response
         """
+        part_number_marker = kwargs.get("part_num_marker", 0)
         if part_number_marker:
-            response = self.s3_client.list_parts(
-                Bucket=bucket, Key=object_name, UploadId=mpu_id,
-                PartNumberMarker=part_number_marker)
+            response = self.s3_client.list_parts(Bucket=bucket_name, Key=object_name,
+                                                 UploadId=mpu_id,
+                                                 PartNumberMarker=part_number_marker)
         else:
-            response = self.s3_client.list_parts(
-                Bucket=bucket, Key=object_name, UploadId=mpu_id)
+            response = self.s3_client.list_parts(Bucket=bucket_name, Key=object_name,
+                                                 UploadId=mpu_id)
         LOGGER.debug(response)
 
         return response
 
-    def complete_multipart_upload(
-            self,
-            mpu_id: str = None,
-            parts: list = None,
-            bucket: str = None,
-            object_name: str = None) -> dict:
+    def complete_multipart_upload(self, mpu_id: str = None, parts: list = None, bucket: str = None,
+                                  object_name: str = None) -> dict:
         """
         Complete a multipart upload, s3 creates an object by concatenating the parts.
 
@@ -129,10 +118,7 @@ class Multipart(S3Lib):
         """
         LOGGER.info("initiated complete multipart upload")
         result = self.s3_client.complete_multipart_upload(
-            Bucket=bucket,
-            Key=object_name,
-            UploadId=mpu_id,
-            MultipartUpload={"Parts": parts})
+            Bucket=bucket, Key=object_name, UploadId=mpu_id, MultipartUpload={"Parts": parts})
         LOGGER.debug(result)
 
         return result
@@ -149,10 +135,8 @@ class Multipart(S3Lib):
 
         return result
 
-    def list_multipart_uploads_with_keymarker(
-            self,
-            bucket: str = None,
-            keymarker: str = None) -> dict:
+    def list_multipart_uploads_with_keymarker(self, bucket: str = None,
+                                              keymarker: str = None) -> dict:
         """
         List all initiated multipart uploads.
         :param bucket: Name of the bucket.
@@ -164,11 +148,8 @@ class Multipart(S3Lib):
 
         return result
 
-    def abort_multipart_upload(
-            self,
-            bucket: str = None,
-            object_name: str = None,
-            upload_id: str = None) -> dict:
+    def abort_multipart_upload(self, bucket: str = None, object_name: str = None,
+                               upload_id: str = None) -> dict:
         """
         Abort multipart upload for given upload_id.
 
@@ -184,11 +165,7 @@ class Multipart(S3Lib):
 
         return response
 
-    def get_object(
-            self,
-            bucket: str = None,
-            key: str = None,
-            ranges: str = None) -> dict:
+    def get_object(self, bucket: str = None, key: str = None, ranges: str = None) -> dict:
         """
         Getting byte range of the object.
 
@@ -197,17 +174,13 @@ class Multipart(S3Lib):
         :param ranges: Range in bytes.
         :return: response.
         """
-        response = self.s3_client.get_object(
-            Bucket=bucket, Key=key, Range=ranges)
+        response = self.s3_client.get_object(Bucket=bucket, Key=key, Range=ranges)
         LOGGER.debug(response)
 
         return response
 
-    def upload_part_copy(self,
-                         copy_source: str = None,
-                         bucket_name: str = None,
-                         object_name: str = None,
-                         **kwargs) -> dict:
+    def upload_part_copy(self, copy_source: str = None, bucket_name: str = None,
+                         object_name: str = None, **kwargs) -> dict:
         """
         Upload parts of a specific multipart upload.
 
@@ -226,31 +199,26 @@ class Multipart(S3Lib):
         if copy_source_range:
             if content_md5:
                 response = self.s3_client.upload_part_copy(
-                    Bucket=bucket_name, Key=object_name,
-                    UploadId=upload_id, PartNumber=part_number,
-                    CopySource=copy_source,
-                    ContentMD5=content_md5,
+                    Bucket=bucket_name, Key=object_name, UploadId=upload_id, PartNumber=part_number,
+                    CopySource=copy_source, ContentMD5=content_md5,
                     CopySourceRange=copy_source_range)
             else:
                 response = self.s3_client.upload_part_copy(
-                    Bucket=bucket_name, Key=object_name,
-                    UploadId=upload_id, PartNumber=part_number,
+                    Bucket=bucket_name, Key=object_name, UploadId=upload_id, PartNumber=part_number,
                     CopySource=copy_source, CopySourceRange=copy_source_range)
         else:
             if content_md5:
                 response = self.s3_client.upload_part_copy(
-                    Bucket=bucket_name, Key=object_name,
-                    UploadId=upload_id, PartNumber=part_number,
-                    CopySource=copy_source,
-                    ContentMD5=content_md5)
+                    Bucket=bucket_name, Key=object_name, UploadId=upload_id, PartNumber=part_number,
+                    CopySource=copy_source, ContentMD5=content_md5)
             else:
                 response = self.s3_client.upload_part_copy(
-                    Bucket=bucket_name, Key=object_name,
-                    UploadId=upload_id, PartNumber=part_number,
+                    Bucket=bucket_name, Key=object_name, UploadId=upload_id, PartNumber=part_number,
                     CopySource=copy_source)
         logging.debug(response)
 
         return response
+
 
 # pylint: disable=too-few-public-methods
 class ProgressPercentage:
@@ -268,10 +236,8 @@ class ProgressPercentage:
         with self._lock:
             self._seen_so_far += bytes_amount
             percentage = (self._seen_so_far / self._size) * 100
-            sys.stdout.write(
-                "\r%s  %s / %s  (%.2f%%)" % (
-                    self._filename, self._seen_so_far, self._size,
-                    percentage))
+            sys.stdout.write("\r%s  %s / %s  (%.2f%%)" % (self._filename, self._seen_so_far,
+                                                          self._size, percentage))
             sys.stdout.flush()
 
 
@@ -301,11 +267,8 @@ class MultipartUsingBoto(S3Lib):
         if s3_prefix is not None:
             key = '/'.join([s3_prefix, key])
         self.s3_resource.Object(bucket_name, key). \
-            upload_file(file_path,
-                        ExtraArgs={'ContentType': 'text/plain'},
-                        Config=config,
-                        Callback=ProgressPercentage(file_path)
-                        )
+            upload_file(file_path, ExtraArgs={'ContentType': 'text/plain'},
+                        Config=config, Callback=ProgressPercentage(file_path))
         return key
 
     def multipart_download(self, **kwargs):
@@ -317,10 +280,7 @@ class MultipartUsingBoto(S3Lib):
         assert key
         try:
             self.s3_resource.Object(bucket_name, key). \
-                download_file(file_path,
-                              Config=config,
-                              Callback=ProgressPercentage(file_path)
-                              )
+                download_file(file_path, Config=config, Callback=ProgressPercentage(file_path))
         except Exception as error:
-            LOGGER.error("Error in multipart_upload: %s", str(error))
+            LOGGER.exception(ERR_MSG, "multipart_download", str(error))
             raise error
